generator logging tools 模块提交

+++ common/src/main/java/com/zhou/example/yadmin/common/utils/FileUtils.java
@@ -89,7 +89,7 @@ public class FileUtils extends org.apache.commons.io.FileUtils
      *
      * @return
      */
-    public static String getSize(int size)
+    public static String getSize(long size)
     {
         return byteCountToDisplaySize(size);
     }
